Moved queue-handling functionality to `items` namespace. Documented funcitonality in readme
