+++ test/unit/items-test.js
@@ -21,5 +21,21 @@ exports.examples = function(expect) {
         });
       });
     });
-	});
+
+    describe('.configureQueue(messageQueueParams)', function() {
+      it('PUTs to /data/v1/items/queues with messageQueueParams as the body', function() {
+        const messageQueueParams = { fromZone: 'exitZone', toZone: 'entryZone' };
+        return expect(this.subject.items.configureQueue(messageQueueParams)).to.wrap.request({
+          method: 'put',
+          path: `/itemsense/data/v1/items/queues`,
+          body: messageQueueParams
+        });
+      });
+    });
+
+    it('handles AMQP events', function() {
+      let AmqpHandler = require('../../dist/services/amqp-handler.service.js');
+      expect(this.subject.health instanceof AmqpHandler.constructor).to.be.true;
+    });
+  });
 }
